refactor: remove boilerplate in benchmarking
